requirements/base.txt
pytz==2023.3.post1  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==10.1.0  # https://github.com/python-pillow/Pillow
rcssmin==1.1.1  # https://github.com/ndparker/rcssmin
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
celery==5.3.6  # pyup: < 6.0  # https://github.com/celery/celery
flower==2.0.1  # https://github.com/mher/flower
xmltodict==0.13.0  # https://github.com/martinblech/xmltodict.git

# Django
# ------------------------------------------------------------------------------
django==5.0.7
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.5.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.63.6  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.3  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.2  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==4.4  # https://github.com/django-compressor/django-compressor
django-redis==5.4.0  # https://github.com/jazzband/django-redis4

# Django REST
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1  # https://django-rest-framework-simplejwt.readthedocs.io/en/latest/

# Wagtail
# ------------------------------------------------------------------------------
wagtail==6.3.2  # https://github.com/wagtail/wagtail

# Wagtail Recaptcha
# ------------------------------------------------------------------------------
django-recaptcha==3.0.0
wagtail-django-recaptcha==1.0

# Wagtail Menu
# ------------------------------------------------------------------------------
wagtailmenus==4.0

# Wagtail Localize
# ------------------------------------------------------------------------------
wagtail-localize==1.9

#Wagtail Admin
# ------------------------------------------------------------------------------
wagtail-modeladmin==2.0.0

# Django celery
# ------------------------------------------------------------------------------
django-celery-beat==2.6.0  # https://github.com/celery/django-celery-beat
django_celery_results==2.5.1

# Wagtail-Autocomplete
# ------------------------------------------------------------------------------
wagtail-autocomplete==0.11.0  # https://github.com/wagtail/wagtail-autocomplete

# Minio
minio==7.2.7

# Reverse Geocode
# ------------------------------------------------------------------------------
reverse-geocode==1.6  # https://pypi.org/project/reverse-geocode/

# SciELO Log Validator
-e git+https://github.com/scieloorg/scielo_log_validator@0.4.0#egg=scielo_log_validator

# SciELO Usage COUNTER
-e git+https://github.com/scieloorg/scielo_usage_counter@1.5.1#egg=scielo_usage_counter

# Device Detector
device-detector==0.10  # https://github.com/thinkwelltwd/device_detector

# langcodes
# ------------------------------------------------------------------------------
langcodes==3.5.0  # https://pypi.org/project/langcodes

# packtools
# ------------------------------------------------------------------------------
lxml==4.9.4 # https://github.com/lxml/lxml
tornado>=6.3.3 # not directly required, pinned by Snyk to avoid a vulnerability
packtools@https://github.com/scieloorg/packtools/archive/refs/tags/3.3.3.zip

# Sickle
# ------------------------------------------------------------------------------
Sickle==0.7.0

# Tenacity
# ------------------------------------------------------------------------------
tenacity==8.3.0  # https://pypi.org/project/tenacity/

# Articlemeta API
# ------------------------------------------------------------------------------
articlemetaapi==1.26.7

# ElasticSearch
# ------------------------------------------------------------------------------
elasticsearch==8.18.1  # https://elasticsearch-py.readthedocs.io/en/v8.18.1/
